pyproject.toml
[project]
name = "cleaning-app"
version = "0.1.0"
description = "A modern, one-page cleaning service booking app built with Flask, HTMX, and Tailwind CSS."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "faker>=37.4.0",
    "flask>=3.1.1",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.1.0",
]
